API

Public interface

I have taken the Observer Pattern as the goal to implement the project. The main objects are Publisher and Subscriber. The Publisher delivers a message to the Subscribers. The information consists of temperatures that were measured.

Exports

API methods

createSubscriber, createPublisher, createMessage, sendMessage

Domain objects

Subscriber, Publisher, Message

SubscriberType values

MEAN_CALCULATOR, STD_CALCULATOR, PLOTTER

PublisherType values

NEWSPAPER, MAGAZINE, SOCIAL_MEDIA

createSubscriber

RbO.createSubscriber — Function.

createSubscriber( name::String )::Sbubscriber

Creates a Subscriber.

Examples:

julia> using RbO

julia> mickey = createSubscriber( "Micky Mouse" )
Subscriber("Micky Mouse", "", MEAN_CALCULATOR::SubscriberType = 0)

source

createSubscriber( name::String, email::String )::Subscriber

Creates a Subscriber with an e-mail address.

Example:

julia> using RbO

julia> daisy = createSubscriber( "Daisy Mouse", "daisy@duckcity.com" )
Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR::SubscriberType = 0)

source

createSubscriber( name::String, email::String, subscribertype::SubscriberType )::Subscriber

Creates a Subscriber with a name, an e-mail address, and SubscriberType.

Example

julia> using RbO

julia> scrooge = createSubscriber( "Scrooge McDuck", "scrooge@duckcity.com", PLOTTER )
Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER::SubscriberType = 2)

source

createPublisher

RbO.createPublisher — Function.

createPublisher( name::String )::Publisher

Creates a Publisher.

Examples

julia> using RbO

julia> nyt = createPublisher("")::Publisher
ERROR: MissingException: Publisher name is mandatory

julia> nyt = createPublisher( "The New York Times" )
Publisher("The New York Times", NEWSPAPER::PublisherType = 0, Subscriber[])

source

createPublisher( name::String, publishertype::PublisherType )::Publisher

Returns a Publisher object.

julia> using RbO

julia> chronicals = createPublisher( "the Duck Chronicals", MAGAZINE )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

source

subscribe

RbO.subscribe — Function.

subscribe( p::Publisher, s::Subscriber )::Publisher

Adds a subscriber to the list with subscribers of a Publisher.

Example

julia> using RbO

julia> daisy = createSubscriber( "Daisy Mouse", "daisy@duckcity.com" )
Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR::SubscriberType = 0)

julia> scrooge = createSubscriber( "Scrooge McDuck", "scrooge@duckcity.com", PLOTTER )
Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER::SubscriberType = 2)

julia> chronicals = createPublisher( "the Duck Chronicals", MAGAZINE )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

julia> subscribe( chronicals, daisy )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR)])

julia> subscribe( chronicals, scrooge )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR), Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER)])

source

unsubscribe

RbO.unsubscribe — Function.

unsubscribe( p::Publisher, s::Subscriber )::Publisher

Removes a subscriber from the list of subscribers of a Publisher,

Example

julia> using RbO

julia> mickey = createSubscriber( "Micky Mouse" )
Subscriber("Micky Mouse", "", MEAN_CALCULATOR::SubscriberType = 0)

julia> chronicals = createPublisher( "the Duck Chronicals", MAGAZINE )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

julia> subscribe( chronicals, mickey )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Micky Mouse", "", MEAN_CALCULATOR)])

julia> chronicals = unsubscribe( chronicals, mickey )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

source

Note

Publisher is an immutable object. 'unscubscribe' creates a new Publisher object, and is reassigned to the variable 'chronicals'.

createMessage

RbO.createMessage — Function.

createMessage(header::String, subject::String, body::Array{Float64, 1})::Message

Create a Message for the subscribers of a publisher.

Example

julia> using RbO

julia> message = createMessage( "Weather station", "Temperatures", [10.9, 12, 10.5, 12.7, 10.2] )
Message("Weather station", "Temperatures", [10.9, 12.0, 10.5, 12.7, 10.2])

source

sendMessage

RbO.sendMessage — Function.

sendMessage( n::Publisher, m::Message, f::Function )

Notifies subscribers

Example

julia> using RbO, Statistics

julia> daisy = createSubscriber( "Daisy Mouse", "daisy@duckcity.com" )
Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR::SubscriberType = 0)

julia> scrooge = createSubscriber( "Scrooge McDuck", "scrooge@duckcity.com", PLOTTER )
Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER::SubscriberType = 2)

julia> chronicals = createPublisher( "the Duck Chronicals", MAGAZINE )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

julia> subscribe( chronicals, daisy )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR)])

julia> subscribe( chronicals, scrooge )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR), Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER)])

julia> message = createMessage( "Weather station", "Temperatures", [10.9, 12, 10.5, 12.7, 10.2] )
Message("Weather station", "Temperatures", [10.9, 12.0, 10.5, 12.7, 10.2])

julia> result = []
0-element Array{Any,1}

julia> processMessage( s::Subscriber, n::Publisher, m::Message ) = begin
           if s.subscribertype == STD_CALCULATOR
               println( s.name * ": the standard deviation of the last five temperatures is: " * string( round( std(m.body); digits=2) ) * " °C" )
           elseif s.subscribertype == MEAN_CALCULATOR
               println( s.name * ": the average of the last five temperatures is: " * string( mean( m.body ) ) * " °C" )
           elseif s.subscribertype == PLOTTER
               println( s.name * ": the dataset with the last five temperatures is: " * string( m.body ) )
               global result = m.body
           end
       end
processMessage (generic function with 1 method)

julia> sendMessage( chronicals, message, processMessage )
Daisy Mouse: the average of the last five temperatures is: 11.26 °C
Scrooge McDuck: the dataset with the last five temperatures is: [10.9, 12.0, 10.5, 12.7, 10.2]

source