User Manual

RbO.jl

Home

RbO.jl

Home

Introduction

Installing the module

Example of adding the module

Example of cloning the module

Public interface

Exports

createSubscriber

createPublisher

subscribe

unsubscribe

createMessage

sendMessage

Domain Items

Subscriber

SubscriberType

Publisher

PublisherType

Message

Examples

test.jl

make.jl

doc.jl

Julia Notebook

References

Introduction

I want to master the Julia programming language. To get more experience, I created the project 'rbontekoe/RbO.jl.'

Also, I decided to use Documenter.jl to create the documentation for the project, after looking at a video that explains its features. I was impressed when Morton Phiibleht told about the possibility to test the code example automatically during the creation of the manual. The result is this manual.

For the creation of the Julia notebook, I used Literate.jl, after watching this video.

Installing the module

The module can be added or cloned from GitHub: https://github.com/rbontekoe/RbO.jl.

Example of adding the module

julia> ]

(v1.1) pkg> add https://github.com/rbontekoe/RbO.jl

(v1.1) pkg> Ctrl-C

julia> using RbO

Example of cloning the module

$ mkdir julia_projects

$ cd julia_projects

$ git clone https://github.com/rbontekoe/RbO.jl.git

$ cd RbO.jl

$ julia

julia> push!(LOAD_PATH, "/home/rob/julia_projects/RbO.jl/src/")

julia> using RbO

Public interface

I have taken the Observer Pattern as the goal to implement the project. The main objects are Publisher and Subscriber. The Publisher delivers a message to the Subscribers. The information consists of temperatures that were measured.

Exports

API methods

createSubscriber, createPublisher, createMessage, sendMessage

Domain objects

Subscriber, Publisher, Message

SubscriberType values

MEAN_CALCULATOR, STD_CALCULATOR, PLOTTER

PublisherType values

NEWSPAPER, MAGAZINE, SOCIAL_MEDIA

createSubscriber

RbO.createSubscriber — Function.

createSubscriber( name::String )::Sbubscriber

Creates a Subscriber.

Examples:

julia> using RbO

julia> mickey = createSubscriber( "Micky Mouse" )
Subscriber("Micky Mouse", "", MEAN_CALCULATOR::SubscriberType = 0)

source

createSubscriber( name::String, email::String )::Subscriber

Creates a Subscriber with an e-mail address.

Example:

julia> using RbO

julia> daisy = createSubscriber( "Daisy Mouse", "daisy@duckcity.com" )
Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR::SubscriberType = 0)

source

createSubscriber( name::String, email::String, subscribertype::SubscriberType )::Subscriber

Creates a Subscriber with a name, an e-mail address, and SubscriberType.

Example

julia> using RbO

julia> scrooge = createSubscriber( "Scrooge McDuck", "scrooge@duckcity.com", PLOTTER )
Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER::SubscriberType = 2)

source

createPublisher

RbO.createPublisher — Function.

createPublisher( name::String )::Publisher

Creates a Publisher.

Examples

julia> using RbO

julia> nyt = createPublisher("")::Publisher
ERROR: MissingException: Publisher name is mandatory

julia> nyt = createPublisher( "The New York Times" )
Publisher("The New York Times", NEWSPAPER::PublisherType = 0, Subscriber[])

source

createPublisher( name::String, publishertype::PublisherType )::Publisher

Returns a Publisher object.

julia> using RbO

julia> chronicals = createPublisher( "the Duck Chronicals", MAGAZINE )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

source

subscribe

RbO.subscribe — Function.

subscribe( p::Publisher, s::Subscriber )::Publisher

Adds a subscriber to the list with subscribers of a Publisher.

Example

julia> using RbO

julia> daisy = createSubscriber( "Daisy Mouse", "daisy@duckcity.com" )
Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR::SubscriberType = 0)

julia> scrooge = createSubscriber( "Scrooge McDuck", "scrooge@duckcity.com", PLOTTER )
Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER::SubscriberType = 2)

julia> chronicals = createPublisher( "the Duck Chronicals", MAGAZINE )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

julia> subscribe( chronicals, daisy )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR)])

julia> subscribe( chronicals, scrooge )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR), Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER)])

source

unsubscribe

RbO.unsubscribe — Function.

unsubscribe( p::Publisher, s::Subscriber )::Publisher

Removes a subscriber from the list of subscribers of a Publisher,

Example

julia> using RbO

julia> mickey = createSubscriber( "Micky Mouse" )
Subscriber("Micky Mouse", "", MEAN_CALCULATOR::SubscriberType = 0)

julia> chronicals = createPublisher( "the Duck Chronicals", MAGAZINE )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

julia> subscribe( chronicals, mickey )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Micky Mouse", "", MEAN_CALCULATOR)])

julia> chronicals = unsubscribe( chronicals, mickey )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

source

Note

Publisher is an immutable object. 'unscubscribe' creates a new Publisher object, and is reassigned to the variable 'chronicals'.

createMessage

RbO.createMessage — Function.

createMessage(header::String, subject::String, body::Array{Float64, 1})::Message

Create a Message for the subscribers of a publisher.

Example

julia> using RbO

julia> message = createMessage( "Weather station", "Temperatures", [10.9, 12, 10.5, 12.7, 10.2] )
Message("Weather station", "Temperatures", [10.9, 12.0, 10.5, 12.7, 10.2])

source

sendMessage

RbO.sendMessage — Function.

sendMessage( n::Publisher, m::Message, f::Function )

Notifies subscribers

Example

julia> using RbO, Statistics

julia> daisy = createSubscriber( "Daisy Mouse", "daisy@duckcity.com" )
Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR::SubscriberType = 0)

julia> scrooge = createSubscriber( "Scrooge McDuck", "scrooge@duckcity.com", PLOTTER )
Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER::SubscriberType = 2)

julia> chronicals = createPublisher( "the Duck Chronicals", MAGAZINE )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[])

julia> subscribe( chronicals, daisy )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR)])

julia> subscribe( chronicals, scrooge )
Publisher("the Duck Chronicals", MAGAZINE::PublisherType = 1, Subscriber[Subscriber("Daisy Mouse", "daisy@duckcity.com", MEAN_CALCULATOR), Subscriber("Scrooge McDuck", "scrooge@duckcity.com", PLOTTER)])

julia> message = createMessage( "Weather station", "Temperatures", [10.9, 12, 10.5, 12.7, 10.2] )
Message("Weather station", "Temperatures", [10.9, 12.0, 10.5, 12.7, 10.2])

julia> result = []
0-element Array{Any,1}

julia> processMessage( s::Subscriber, n::Publisher, m::Message ) = begin
           if s.subscribertype == STD_CALCULATOR
               println( s.name * ": the standard deviation of the last five temperatures is: " * string( round( std(m.body); digits=2) ) * " °C" )
           elseif s.subscribertype == MEAN_CALCULATOR
               println( s.name * ": the average of the last five temperatures is: " * string( mean( m.body ) ) * " °C" )
           elseif s.subscribertype == PLOTTER
               println( s.name * ": the dataset with the last five temperatures is: " * string( m.body ) )
               global result = m.body
           end
       end
processMessage (generic function with 1 method)

julia> sendMessage( chronicals, message, processMessage )
Daisy Mouse: the average of the last five temperatures is: 11.26 °C
Scrooge McDuck: the dataset with the last five temperatures is: [10.9, 12.0, 10.5, 12.7, 10.2]

source

Domain Items

Subscriber

RbO.Subscriber — Type.

struct Subscriber

Returns a Subscriber. It is preferred to use createSubscriber.

Example

julia> using RbO

julia> daisy = Subscriber("Daisy")
Subscriber("Daisy", "", MEAN_CALCULATOR::SubscriberType = 0)

source

SubscriberType

RbO.SubscriberType — Type.

SubscriberType

Values

STD_CALCULATOR (caluclates the standard deviation of set of data, default value)

MEAN_CALCULATOR (caluclates the mean of set of data)

PLOTTER (keeps the dataset for plotting)

source

Publisher

RbO.Publisher — Type.

struct Publisher

Returns a Publisher object. It is preferred to use createPublisher.

Example

julia> using RbO

julia> chronicals = Publisher("The Duck City Chornicals", NEWSPAPER, Subscriber[])
Publisher("The Duck City Chornicals", NEWSPAPER::PublisherType = 0, Subscriber[])

source

PublisherType

RbO.PublisherType — Type.

PublisherType

Values

NEWSPAPER

MAGAZINE

SOCIAL_MEDIA

source

Message

RbO.Message — Type.

struct Message

Returms a Message object. It is preferred to use createMessage.

Example

julia> using RbO

julia> message = Message( "Weather station", "Temperatures", [10.9, 12, 10.5, 12.7, 10.2] )
Message("Weather station", "Temperatures", [10.9, 12.0, 10.5, 12.7, 10.2])

source

Examples

test.jl

Code to test RbO

using RbO, Plots, Statistics
pyplot()

# ### Create subscribers
daisy = createSubscriber( "Daisy Duck", "daisy@duckcity.com" )
donald = createSubscriber( "Donald Duck", "donald@duckcity.com", STD_CALCULATOR )
scrooge = createSubscriber( "Scrooge McDuck", "scrooge@duckcity.com", PLOTTER )
mickey = createSubscriber( "Micky Mouse" )

# ### Create newspapers and magazines
chronicals = createPublisher( "the Duck Chronicals" )

# Add subscribers to the Duck Chronicals
subscribe( chronicals, daisy )
subscribe( chronicals, donald )
subscribe( chronicals, scrooge )
subscribe( chronicals, mickey )

# Unsubscribe Mickey Mouse
chronicals = unsubscribe(chronicals, mickey)

# ### Define message for the Duck Chronical subscribers
message = createMessage( "Weather station", "Temperatures", [10.9, 12, 10.5, 12.7, 10.2] )

# ### Print the messages
paper = chronicals

# List with the subscribers of a newspaper or magazine
sm = map( x -> x.email == "" ? x.name : x.name * " - " * x.email, paper.list )

# Print the subscribers
len =  length( "Subscribers of " * paper.name * string(paper.publishertype )) + 5

println( "\n" * "="^len )
title = titlecase( string( paper.publishertype ) )
println( " Subscribers of " * paper.name * " - " * map(x -> x == '_' ? ' ' : x , title ) )
println( "="^len )
for subscriber in sm
    println( " " * subscriber )
end
println( '='^len )

# ### Define the function 'processMessage' to process data depending on SubsriberType
#result = []
processMessage( s::Subscriber, n::Publisher, m::Message ) = begin
    if s.subscribertype == STD_CALCULATOR
        println( s.name * ":\t the standard deviation of the last five temperatures is: " * string( round( std(m.body); digits=2) ) * " °C" )
    elseif s.subscribertype == MEAN_CALCULATOR
        #println( s.name * ":\t the average of the last five temperatures is: " * string(sum(m.body) / length(m.body)) * " °C" )
        println( s.name * ":\t the average of the last five temperatures is: " * string( mean( m.body ) ) * " °C" )
    elseif s.subscribertype == PLOTTER
        println( s.name * ":\t the dataset with the last five temperatures is: " * string( m.body ) )
        global result = m.body
    end
end

# ### Notify scubscribers
sendMessage( chronicals, message, processMessage )

# ### Plot data
plot( result )

make.jl

using Documenter
using RbO, Statistics
using Test

doctest(RbO)

makedocs(
    sitename = "RbO",
    format = Documenter.HTML(),
    modules = [RbO],
    pages = [
        "User Manual" => "index.md",
        "Building applications with Julia" => "course.md"
    ]
)

doc.jl

Code to create the Julia Notebook

# Uncomment the next line when RbO has been cloned.
#push!(LOAD_PATH, "/home/rob/julia_projects/RbO.jl/src/")
include("./test.jl")
using Literate
Literate.notebook("./test.jl", "."; documenter=false)

Julia Notebook

You can find an example of the code in a Julia Notebook on GitHub. See: https://github.com/rbontekoe/RbO.jl/blob/master/test.ipynb

References

Parallel & Distributed Computing